Sort dNdS Values Into Divergence Strata

This function takes a data.table returned by dNdS and sorts the corresponding dNdS value into divergence strata (deciles).

DivergenceMap(dNdS_tbl, subject.id = FALSE)

Arguments

dNdS_tbl
a data.table object returned by dNdS.

subject.id
a logical value indicating whether query_id AND subject_id should be returned.

Value

a data.table storing a standard divergence map.

Details

Divergence Strata are decile values of corresponding dNdS values. The dNdS function returns dNdS values for orthologous genes of a query species (versus subject species). These dNdS values are then sorted into deciles and each orthologous protein coding gene of the query species receives a corresponding decile value instead of the initial dNdS value.

This allows a better comparison between Phylostrata and Divergence Strata (for more details see package: myTAI).

See also

Examples

# NOT RUN {
# get a divergence map of example sequences
dNdS_tbl <- dNdS(
              query_file      = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
              subject_file    = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
              ortho_detection = "RBH",
              aa_aln_type     = "pairwise",
              aa_aln_tool     = "NW",
              codon_aln_tool  = "pal2nal",
              dnds_est.method = "Comeron",
              comp_cores      = 1 )

divMap <- DivergenceMap( dNdS_tbl )

# in case you want the subject_id as well, you can set
# the argument subject.id = TRUE
divMap <- DivergenceMap( dNdS_tbl = dNdS_tbl, subject.id = TRUE)

# }