Infer orthologous lncRNAs between muliple species

Inference of orthologous lncRNAs between multiple species is performed via pairwise comparisons. The corresponding orthologous tables are then stored in an output folder.

map.generator.lnc(query_file, subjects.folder, output.folder,
  eval = "1E-5", ortho_detection = "RBH", comp_cores = 1,
  progress.bar = TRUE, sep = ";", ...)

Arguments

query_file
a character string specifying the path to the lncRNAs file of the query organism in fasta format.

subjects.folder
a character string specifying the path to the folder where lncRNAs files in fasta format of the subject organisms are stored.

output.folder
a character string specifying the path to the folder where output orthologous tables should be stored.

eval
a character string specifying the e-value for BLAST based orthology inference. Please use the scientific notation.

ortho_detection
a character string specifying the Orthology Inference method that shall be used to perform dNdS computations. Possible options are:

ortho_detection = "BH": BLAST best unidirectional hit

ortho_detection = "RBH": BLAST best reciprocal hit

comp_cores
number of computing cores that shall be used to perform parallelized computations.

progress.bar
should a progress bar be shown. Default is progress.bar = TRUE.

sep
a file separator that is used to store maps as csv file.

...
additional parameters that shall be passed to dNdS.

Details

Given a query organism and a set of subject organsisms that are stored in the same folder, this function crawls through all subject organsism and infers the lncRNA homologs in pairwise species comparisons.

Examples

# NOT RUN {
map.generator.lnc(
   query_file      = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
   subjects.folder = system.file('seqs/map_gen_example', package = 'orthologr'),
   output.folder   = getwd(),
   quiet           = TRUE,
   comp_cores      = 1
)
# }