Interface function to Orthofinder2

Run Orthofinder2 from R.

orthofinder2(
  proteome_folder,
  use_existing_output = FALSE,
  import_type = "orthogroups_core",
  comp_cores = 1
)

Arguments

proteome_folder
file path to a folder storing the proteome sequences of the species for which orthology inference shall be performed.

use_existing_output
a logical value indicating whether or not an existing Orthofinder2 output folder shall be used fo further import and processing. If use_existing_output = TRUE is selected then please specify the file path to the in proteome_folder where the Orthofinder2 output folder can be found.

import_type
type of Orthofinder2 output that shall be imported after running Orthofinder2. Options are:

import_type = "orthogroups_core"

import_type = "orthogroups_pairwise"

import_type = ""

import_type = ""

comp_cores
number of cores that shall be used for parallel processing. Default is cores = 1.

Examples

if (FALSE) {
# specify species names
orgs <- c("Arabidopsis lyrata",
          "Capsella rubella", "Solanum lycopersicum")
# download proteome files for all species          
biomartr::getProteomeSet(db = "refseq", organisms = orgs, path = "of_proteomes")
# download annotation files for all species          
biomartr::getGFFSet(db = "refseq", organisms = orgs, path = "of_gff")
# select longest splice variant per gene locus
retrieve_longest_isoforms_all(proteome_folder = "of_proteomes",
                              annotation_folder = "of_gff",
                              annotation_format = "gff",
                              output_folder = "of_proteomes_longest_sv")
# run orthofinder2 to infer ortho groups for the specified species
orthofinder2(proteome_folder = "of_proteomes_longest_sv", comp_cores = 4)
}